refactor: abstract replace_nan
