+++ kolena/_utils/dataframes/__init__.py
@@ -11,6 +11,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from kolena._utils.dataframes._pandas import df_apply
-
-__all__ = [df_apply]
